* kadm_funcs.c: Cast argument to isalnum()/isdigit() to int
	* acl_files.c (nuke_whitespace): Cast argument to isspace() to int.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13375 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/kadmin/v4server/acl_files.c
@@ -280,7 +280,7 @@ char *buf;
     register char *pin, *pout;
 
     for(pin = pout = buf; *pin != '\0'; pin++)
-	if(!isspace(*pin)) *pout++ = *pin;
+	if(!isspace((int) *pin)) *pout++ = *pin;
     *pout = '\0';		/* Terminate the string */
 }
